Fix terrain Balance hard and Add Main Menu button
Fix the mountain terrain on balance hard 2 and Add main menu button on UI prefab

fixed the mountain level on balance 2, as the terrain did not load. also added a button in the lower left corner of the level UI to go back to the main menu to help with testing.